convert int to uint8

+++ message/codec_test.go
@@ -114,7 +114,7 @@ func TestCodecPackParseGzip(t *testing.T) {
 		{
 			op: UptimePong,
 			fields: map[Field]interface{}{
-				Uptime: 80,
+				Uptime: uint8(80),
 			},
 		},
 		{
